+++ package/Editor/Extension/UI/UIUtils.cs
@@ -1,4 +1,8 @@
-﻿using UnityEngine.Experimental.UIElements;
+﻿#if UNITY_2019_1_OR_NEWER
+	using UnityEngine.UIElements;
+#else
+	using UnityEngine.Experimental.UIElements;
+#endif
 
 namespace UnityEditor.PackageManager.ValidationSuite.UI
 {